🚧 Slight rearrage, move plotting away from config
