--- bluemira/power_cycle/net/manager.py
@@ -1,1 +1,0 @@
-
